feat(univention_directory_manager): try to modify objects without dn
Previously every object without dn was created which led to failures
upon subsequent runs. Now if a uid or name is present a corresponding
object is queried via get_by_id().
This will still fail if either of them are not given or the object is
present but cannot be found.