Univention Config Registry: manually read registry files
Until this change Univention's own config registry backend module was
loaded & its functionality used for reading the UCR. This is
unfortunately incompatible with using the Mitogen connection strategy
plugin & Python 3.

Mitogen works by streaming Python modules from the Ansible host to the
target machine, meaning the imported modules must exist on the host
machine, too, which would limit the Ansible host to being a Univention
system, too.

Interestingly this only manifests when Python 3 is used as the Python
interpreter on the target; it works fine with Python 2.

This change does two things:

1. It reads the config registry files directly. As the config registry
   is distributed over different files with different priorities, the
   routine follows Univention's backend module wrt. to the
   priority/order in which the files are loaded: settings from files
   with higher priority always overwrite the same settings from files
   with lower priority.

2. Committing templates isn't done via the backend module
   anymore. Instead the external "ucr commit" command is used much
   like it is for setting or unsetting keys.